refactor: shared/ui 폴더 구조 정리 및 컴포넌트 분리
