+++ README.md
@@ -81,3 +81,8 @@ TanstackQuery를 이용하여 코드를 개선하기
 - [ ] fetch와 useState가 아닌 선언적인 함수형 프로그래밍이 적절히 적용되었는가?
 - [ ] 캐싱과 리프레시 전략이 올바르게 구현되었는가?
 ```
+
+**단계별 진행 과정**
+1. 상태관리 설치
+2. 공통 컴포넌트 UI 분리
+2.
